deprecate Formatters module
